adjusted folder structure
