+++ src/pages/about/about-page.js
@@ -1,6 +1,7 @@
 import { Html, Scroll, ScrollControls } from "@react-three/drei";
 import { useThree } from "@react-three/fiber";
 import { animated, config, useSpring } from "@react-spring/web";
+import { CameraView } from "../../routing-test/camera-view";
 
 export const AboutPage = () => {
     const { width, height } = useThree((state) => state.size);
@@ -12,85 +13,90 @@ export const AboutPage = () => {
     });
 
     return (
-        <MobileScrollWrapper width={width} height={height}>
-            <Html fullscreen position={[0, -2.2, 0]}>
-                <animated.div
-                    className="body"
-                    style={{
-                        display: "flex",
-                        opacity,
-                        flexDirection: width > 1000 ? "row" : "column",
-                        justifyContent: "space-between",
-                        textAlign: "left",
-                    }}
-                >
-                    <div
+        <CameraView>
+            <MobileScrollWrapper width={width} height={height}>
+                <Html fullscreen position={[0, -2.2, 0]}>
+                    <animated.div
+                        className="body"
                         style={{
                             display: "flex",
-                            flexDirection: "column",
-                            justifyContent: "center",
-                            width: width > 1000 ? "50%" : "100%",
-                            marginRight:
-                                width > 1000 && height > 1000 ? "0px" : "100px",
+                            opacity,
+                            flexDirection: width > 1000 ? "row" : "column",
+                            justifyContent: "space-between",
+                            textAlign: "left",
                         }}
                     >
-                        <h1
-                            style={{
-                                fontFamily: "Gilroy",
-                                fontSize:
-                                    width > 1000
-                                        ? "7em"
-                                        : width < 400
-                                        ? "2em"
-                                        : "4em",
-                                fontWeight: "bold",
-                            }}
-                        >
-                            ABOUT ME
-                        </h1>
-
                         <div
                             style={{
-                                fontFamily: "Gilroy",
-                                fontSize: "1.8em",
-                                fontWeight: "bold",
+                                display: "flex",
+                                flexDirection: "column",
+                                justifyContent: "center",
+                                width: width > 1000 ? "50%" : "100%",
+                                marginRight:
+                                    width > 1000 && height > 1000
+                                        ? "0px"
+                                        : "100px",
                             }}
                         >
-                            <p>
-                                I'm a creative developer working on apps and web
-                                projects all the time. I started this journey
-                                when I was 16 years old and have been working
-                                full time as a developer since 2018.
-                            </p>
-                            <p>
-                                Besides that, for almost the same amount of
-                                time, I've been a photographer. Starting with
-                                just my phone in 2019 I worked my way up to a
-                                professional camera and now even doing paid work
-                                for clients sometimes.
-                            </p>
+                            <h1
+                                style={{
+                                    fontFamily: "Gilroy",
+                                    fontSize:
+                                        width > 1000
+                                            ? "7em"
+                                            : width < 400
+                                            ? "2em"
+                                            : "4em",
+                                    fontWeight: "bold",
+                                }}
+                            >
+                                ABOUT ME
+                            </h1>
+
+                            <div
+                                style={{
+                                    fontFamily: "Gilroy",
+                                    fontSize: "1.8em",
+                                    fontWeight: "bold",
+                                }}
+                            >
+                                <p>
+                                    I'm a creative developer working on apps and
+                                    web projects all the time. I started this
+                                    journey when I was 16 years old and have
+                                    been working full time as a developer since
+                                    2018.
+                                </p>
+                                <p>
+                                    Besides that, for almost the same amount of
+                                    time, I've been a photographer. Starting
+                                    with just my phone in 2019 I worked my way
+                                    up to a professional camera and now even
+                                    doing paid work for clients sometimes.
+                                </p>
+                            </div>
                         </div>
-                    </div>
-                    <div
-                        style={{
-                            paddingTop: "5em",
-                            height:
-                                width > 1000 && height > 1000
-                                    ? "800px"
-                                    : "100%",
-                        }}
-                    >
-                        <img
-                            src="portrait.jpeg"
+                        <div
                             style={{
-                                maxWidth: "100%",
-                                maxHeight: "100%",
+                                paddingTop: "5em",
+                                height:
+                                    width > 1000 && height > 1000
+                                        ? "800px"
+                                        : "100%",
                             }}
-                        />
-                    </div>
-                </animated.div>
-            </Html>
-        </MobileScrollWrapper>
+                        >
+                            <img
+                                src="portrait.jpeg"
+                                style={{
+                                    maxWidth: "100%",
+                                    maxHeight: "100%",
+                                }}
+                            />
+                        </div>
+                    </animated.div>
+                </Html>
+            </MobileScrollWrapper>
+        </CameraView>
     );
 };
